Move layers to separate projects
