+++ src/tasks-api/BusinessLayer/Ekids.Chatbot.Tasks.BusinessLayer.csproj
@@ -1,0 +1,13 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\DataAccessLayer\EKids.Chatbot.Tasks.DataAccessLayer.csproj" />
+  </ItemGroup>
+
+</Project>
