+++ src/tasks-api/BusinessLayer/ILearningTaskService.cs
@@ -1,6 +1,9 @@
 ﻿using EKids.Chatbot.Tasks.DataAccessLayer.Entities;
 
 namespace EKids.Chatbot.Tasks.BusinessLayer;
+/// <summary>
+/// Defines business layer of CRUD methods for managing learning tasks associated with courses.
+/// </summary>
 public interface ILearningTaskService
 {
     Task AddTask(LearningTask task, Guid courseId, CancellationToken cancellation);
