+++ src/tasks-api/DataAccessLayer/CosmosSystemTextJsonSerializer.cs
@@ -1,5 +1,6 @@
 ﻿#nullable disable
 using Microsoft.Azure.Cosmos;
+using System.IO;
 using System.Text.Json;
 
 namespace EKids.Chatbot.Tasks.DataAccessLayer;
