+++ src/tasks-api/DataAccessLayer/Entities/TaskComplexity.cs
@@ -1,4 +1,5 @@
-﻿using System.Text.Json.Serialization;
+﻿using System;
+using System.Text.Json.Serialization;
 
 namespace EKids.Chatbot.Tasks.DataAccessLayer.Entities;
