+++ src/tasks-api/DataAccessLayer/ILearningTaskRepository.cs
@@ -1,6 +1,13 @@
 ﻿using EKids.Chatbot.Tasks.DataAccessLayer.Entities;
+using System;
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace EKids.Chatbot.Tasks.DataAccessLayer;
+/// <summary>
+/// Defines data access layer of CRUD methods for managing learning tasks associated with courses.
+/// </summary>
 public interface ILearningTaskRepository
 {
     Task Save(IEnumerable<LearningTask> tasks, CancellationToken cancellation);
