+++ src/tasks-api/EKids.Chatbot.Tasks.sln
@@ -3,7 +3,19 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.9.34728.123
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EKids.Chatbot.Tasks", "EKids.Chatbot.Tasks.csproj", "{9B9AC2F4-2BC5-4B9D-B74C-456C5886457E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EKids.Chatbot.Tasks.WebApi", "PresentationLayer\EKids.Chatbot.Tasks.WebApi.csproj", "{9B9AC2F4-2BC5-4B9D-B74C-456C5886457E}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EKids.Chatbot.Tasks.BusinessLayer", "BusinessLayer\EKids.Chatbot.Tasks.BusinessLayer.csproj", "{365EFE40-D2FA-47EC-9093-D1CF53A771AC}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EKids.Chatbot.Tasks.DataAccessLayer", "DataAccessLayer\EKids.Chatbot.Tasks.DataAccessLayer.csproj", "{1EB3EE84-9A1A-4A6F-BA03-56803912ADCB}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "PresentationLayer", "PresentationLayer", "{BC45B057-7135-4B3E-A5D9-0D1D7E31C315}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "BusinessLayer", "BusinessLayer", "{176BF050-1AF7-4CFB-BB0C-E9B9D627C473}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "DataAccessLayer", "DataAccessLayer", "{1D04D457-CA5A-435F-9D60-FA5B394010D2}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "EKids.Chatbot.Tasks.WebApi.Tests", "tests\EKids.Chatbot.Tasks.WebApi.Tests\EKids.Chatbot.Tasks.WebApi.Tests.csproj", "{807F6C7E-0C36-49CB-B2C7-32471BF05F38}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,9 +27,27 @@ Global
 		{9B9AC2F4-2BC5-4B9D-B74C-456C5886457E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9B9AC2F4-2BC5-4B9D-B74C-456C5886457E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9B9AC2F4-2BC5-4B9D-B74C-456C5886457E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{365EFE40-D2FA-47EC-9093-D1CF53A771AC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{365EFE40-D2FA-47EC-9093-D1CF53A771AC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{365EFE40-D2FA-47EC-9093-D1CF53A771AC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{365EFE40-D2FA-47EC-9093-D1CF53A771AC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1EB3EE84-9A1A-4A6F-BA03-56803912ADCB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1EB3EE84-9A1A-4A6F-BA03-56803912ADCB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1EB3EE84-9A1A-4A6F-BA03-56803912ADCB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1EB3EE84-9A1A-4A6F-BA03-56803912ADCB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{807F6C7E-0C36-49CB-B2C7-32471BF05F38}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{807F6C7E-0C36-49CB-B2C7-32471BF05F38}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{807F6C7E-0C36-49CB-B2C7-32471BF05F38}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{807F6C7E-0C36-49CB-B2C7-32471BF05F38}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{9B9AC2F4-2BC5-4B9D-B74C-456C5886457E} = {BC45B057-7135-4B3E-A5D9-0D1D7E31C315}
+		{365EFE40-D2FA-47EC-9093-D1CF53A771AC} = {176BF050-1AF7-4CFB-BB0C-E9B9D627C473}
+		{1EB3EE84-9A1A-4A6F-BA03-56803912ADCB} = {1D04D457-CA5A-435F-9D60-FA5B394010D2}
+		{807F6C7E-0C36-49CB-B2C7-32471BF05F38} = {BC45B057-7135-4B3E-A5D9-0D1D7E31C315}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {F2160B78-1E84-42E8-9640-10E42260FCC2}
